Fetaure/transfer

+++ backend/database/migrations/2023_01_06_015024_create_transactions_table.php
@@ -16,7 +16,7 @@ class CreateTransactionsTable extends Migration
         Schema::create('transactions', function (Blueprint $table) {
             $table->id();
             $table->integer('code');
-            $table->integer('amount');
+            $table->bigInteger('amount');
             $table->unsignedInteger('from_id')->nullable()->index();
             $table->unsignedInteger('to_id')->nullable()->index();
             $table->string('message');
